+++ src/components/DeprecatedUnpackedItem/UnpackedItem.css
@@ -1,0 +1,81 @@
+.item--unpacked {
+  background-color: white;
+  box-shadow: 0 1px 2px 0 rgba(0, 0, 0, 0.2),
+    0 1.5px 3.33px 0 rgba(0, 0, 0, 0.19);
+  width: 60px;
+  height: 102px;
+}
+
+.item--count--packed {
+  font-size: 7px;
+  color: white;
+}
+
+.item--count--size {
+  font-size: 7px;
+}
+
+.item--icon {
+  text-align: right;
+}
+
+.item--icon--filler {
+  height: 24px;
+}
+
+.item--icon--packed {
+  font-size: 12px;
+  color: white;
+}
+
+.item--icon--size {
+  font-size: 12px;
+  color: #454ade;
+}
+
+.item--image--size {
+  width: 40px;
+  height: 40px;
+}
+
+.item--input--count {
+  width: 15px;
+  height: 13px;
+  font-size: 10px;
+  font-weight: 350;
+}
+
+.item--pack--font {
+  font-size: 8px;
+  font-weight: 500;
+}
+
+.item--text--selected {
+  border: 1px solid #454ade;
+  background-color: #454ade;
+  color: white;
+}
+
+.item--text--unselected {
+  border: 1px solid #454ade;
+  background-color: white;
+  color: #454ade;
+}
+
+.item--title--font {
+  font-size: 10px;
+}
+
+.item--title--color {
+  color: #454ade;
+}
+
+.item--icon-filler {
+  height: 24px;
+}
+
+.item--description-flow {
+  height: 14px;
+  overflow-x: scroll;
+  overflow-y: auto;
+}
